Move feature calculation to own file
