+++ ctapipe/calib/array/muon/__init__.py
@@ -1,4 +1,5 @@
 from .fitting import kundu_chaudhuri_circle_fit
 from .fitting import psf_likelihood_fit
-from .fitting import mean_squared_error
-from .fitting import photon_ratio_inside_ring
+
+from .features import mean_squared_error
+from .features import photon_ratio_inside_ring
